split editor setup logic to monaco directory
